+++ src/config.py
@@ -79,6 +79,10 @@ def is_ipv6_test_enabled():
     return parse_dlc_developer_configs("test", "enable_ipv6")
 
 
+def is_new_test_structure_enabled():
+    return parse_dlc_developer_configs("test", "use_new_test_structure")
+
+
 def is_ecs_test_enabled():
     return parse_dlc_developer_configs("test", "ecs_tests")
 
@@ -117,10 +121,6 @@ def is_sm_local_test_enabled():
 
 def is_nightly_pr_test_mode_enabled():
     return parse_dlc_developer_configs("test", "nightly_pr_test_mode")
-
-
-def is_scheduler_enabled():
-    return parse_dlc_developer_configs("test", "use_scheduler")
 
 
 def is_safety_check_test_enabled():