+++ vllm/CHANGELOG.md
@@ -1,6 +1,66 @@
 # Changelog
 
-All notable changes to vLLM Deep Learning Containers will be documented in this file.
+Available in Amazon ECR Public Gallery: https://gallery.ecr.aws/deep-learning-containers/vllm.
+
+All notable changes to vLLM Deep Learning Containers will be documented in this file. 
+
+## [0.12.0] - 2025-12-08
+### Added
+- SageMaker and EC2 support for vLLM 0.12.0
+- Upgraded EFA version to 1.45.1 and introduced NIXL with Libfabric backend
+
+### Sample ECR URI
+```
+public.ecr.aws/deep-learning-containers/vllm:0.12.0-gpu-py312
+public.ecr.aws/deep-learning-containers/vllm:0.12.0-gpu-py312-ec2
+```
+
+## [0.11.2] - 2025-11-20
+### Added
+- SageMaker and EC2 support for vLLM 0.11.2
+- CUDA 12.9 and Python 3.12 compatibility
+- Latest ML performance optimizations inherited from vLLM 0.11.2
+
+### Sample ECR URI
+```
+public.ecr.aws/deep-learning-containers/vllm:0.11.2-gpu-py312
+public.ecr.aws/deep-learning-containers/vllm:0.11.2-gpu-py312-ec2
+```
+
+## [0.11.1] - 2025-11-19
+### Added
+- SageMaker and EC2 support for vLLM 0.11.1
+- CUDA 12.9 and Python 3.12 compatibility
+- Latest ML performance optimizations inherited from vLLM 0.11.1 
+
+### Sample ECR URI
+```
+public.ecr.aws/deep-learning-containers/vllm:0.11.1-gpu-py312
+public.ecr.aws/deep-learning-containers/vllm:0.11.1-gpu-py312-ec2
+```
+
+## [0.11.0] - 2025-10-14
+### Added
+- SageMaker support for vLLM 0.11.0
+- NIXL 0.6.0 and LMCache 0.3.7 libraries included for faster model inference performance
+
+### Sample ECR URI
+```
+763104351884.dkr.ecr.us-east-1.amazonaws.com/vllm:0.11.0-gpu-py312-cu128-ubuntu22.04-sagemaker-v1.0
+763104351884.dkr.ecr.us-east-1.amazonaws.com/vllm:0.11.0-gpu-py312-cu128-ubuntu22.04-sagemaker
+```
+
+## [0.11.0] - 2025-10-08
+### Updated
+- vllm/vllm-openai version `v0.11.0`, see [release note](https://github.com/vllm-project/vllm/releases/tag/v0.11.0) for details.
+- EFA installer version `1.43.3`
+
+
+### Sample ECR URI
+```
+763104351884.dkr.ecr.us-east-1.amazonaws.com/vllm:0.11-gpu-py312-ec2
+763104351884.dkr.ecr.us-east-1.amazonaws.com/vllm:0.11.0-gpu-py312-cu128-ubuntu22.04-ec2
+```
 
 ## [0.10.2] - 2025-09-18
 ### Updated